audit

+++ src/subcmd/hmmscan.rs
@@ -49,7 +49,7 @@ struct Args {
 
     /// Report profiles >= this score threshold
     #[arg(short = 'T', conflicts_with = "e_value", allow_hyphen_values = true)]
-    score_threshold: Option<f32>,
+    score_threshold: Option<f64>,
 
     /// Report domains <= this E-value threshold
     #[arg(
@@ -62,7 +62,7 @@ struct Args {
 
     /// Report domains >= this score threshold
     #[arg(long = "domT", conflicts_with = "dom_e", allow_hyphen_values = true)]
-    dom_t: Option<f32>,
+    dom_t: Option<f64>,
 
     /// Include profiles <= this E-value threshold
     #[arg(
@@ -75,7 +75,7 @@ struct Args {
 
     /// Include profiles >= this score threshold
     #[arg(long = "incT", conflicts_with = "inc_e", allow_hyphen_values = true)]
-    inc_t: Option<f32>,
+    inc_t: Option<f64>,
 
     /// Include domains <= this E-value threshold
     #[arg(
@@ -88,7 +88,7 @@ struct Args {
 
     /// Include domains >= this score threshold
     #[arg(long = "incdomT", conflicts_with = "inc_dome", allow_hyphen_values = true)]
-    inc_dom_t: Option<f32>,
+    inc_dom_t: Option<f64>,
 
     /// Use model's GA gathering cutoffs to set all thresholding
     #[arg(
